Update installation.md
I spent hours trying to solve this and nearly gave up on this plugin because of it 😅. When I clicked "Add Sketch" nothing would happen.

I only found out about this step when I scrolled through recent updates on Discord. This PR adds the information to the installation doc.

I also suggest adding it to the instructions on the Gumroad page, but I'm not sure that can be done from this repo.