Add context manager for safe constraints
You add the constraints as normal, and when the context manager exits it checks if the sketch still solves.
If it doesn't, it removes the last constraint and checks again.
This continues until either the sketch solves, or all the constraints that were added inside the context manager are removed.
The fact that the loop checks whether the sketch solves one more time after the list of new constraints is empty is deliberate; it makes sure the sketch is up to date.
The loop doesn't check whether the list is empty, because when it's empty the sketch should solve (hence the check to make sure it solved before any of the constraints were added).