Mesh extrude, Boolean mode
Added an option to 'Extrude' a 3rd dimension to sketches when converting to meshes. 
This adds a Solidify modifier automatically and fills in the 'Depth' and 'Offset' from sketch properties `extrude_depth` and `extrude_offset`
It first checks for and adjusts the existing Solidify modifier, and creates one if it doesn't exist.

Added a 4th conversion mode to the `update_convertor_geometry` function - Boolean
This will convert to mesh, and allows for the new 'Extrude', but also sets the viewport display to 'Bounds' and allows you to specify the target object you wish to cut, automatically applying a Boolean modifier to the target, and selecting your current sketch as the cutter object.