Checks for duplicate constraints. (#461)
* Checks for duplicate constraints.
On constraint creation checks are added to see if it exists on the selected entity or entitiies.
The exists method was added to GenericConstraintOp for this.
Line distances have been replaced for point to point distances.

* Flake8 cleanups

* Simplify exists method and version changes
Simplified the exists method in base_constraint not to do specific checks and moved this logic into the add_distance operator.
Added version check for older drawings. This updates all distance constraints on a line to the endpoints of that line.
Changed addon version from 0.27.3 to 0.27.4

* Removed nested if statements
Added extra check on length of line_entities to avoid errors if it would somehow be out of bounds